+++ packages/elixir/checksum-Elixir.LiterLlm.Native.exs
@@ -1,10 +1,10 @@
 %{
-  "libliter_llm_nif-v1.9.0-nif-2.16-aarch64-apple-darwin.so.tar.gz" => "sha256:33093042a23bdc9c923ae8cab4bc3110239615f732a6f22633f5cfce44890a23",
-  "libliter_llm_nif-v1.9.0-nif-2.16-aarch64-unknown-linux-gnu.so.tar.gz" => "sha256:2cfaf49ef118dc0e95c7fa1745122dfd60e6ecc7ad506e1c4a21406f988ef994",
-  "libliter_llm_nif-v1.9.0-nif-2.16-x86_64-pc-windows-gnu.dll.tar.gz" => "sha256:12db7feacbb0f6317000bc2700c31c24bd369154ac9d6c01053a3a71a1f69da9",
-  "libliter_llm_nif-v1.9.0-nif-2.16-x86_64-unknown-linux-gnu.so.tar.gz" => "sha256:10f7ab4ef1200d645039b6c5c91d5488986182c0be655bdae28da47d5841167c",
-  "libliter_llm_nif-v1.9.0-nif-2.17-aarch64-apple-darwin.so.tar.gz" => "sha256:81f22e967bdc06c36ead2198f2f9d891689001d272d3a2fdf30d1340ded60f8a",
-  "libliter_llm_nif-v1.9.0-nif-2.17-aarch64-unknown-linux-gnu.so.tar.gz" => "sha256:e7da767291cd4c3f07afca5779ae4e6c7c174a2fc7b6662e4f7b2c1c8355373d",
-  "libliter_llm_nif-v1.9.0-nif-2.17-x86_64-pc-windows-gnu.dll.tar.gz" => "sha256:4e7fafb6356b839d9cbfffe7a76f7a84e1f84121e92dafbe5dcc0b26b32c19ac",
-  "libliter_llm_nif-v1.9.0-nif-2.17-x86_64-unknown-linux-gnu.so.tar.gz" => "sha256:8395235401094ac18a54114b9a3f27fea36eba82353da8e166c514cb72226865",
+"libliter_llm_nif-v1.9.0-nif-2.16-aarch64-apple-darwin.so.tar.gz" => "sha256:33093042a23bdc9c923ae8cab4bc3110239615f732a6f22633f5cfce44890a23",
+"libliter_llm_nif-v1.9.0-nif-2.16-aarch64-unknown-linux-gnu.so.tar.gz" => "sha256:2cfaf49ef118dc0e95c7fa1745122dfd60e6ecc7ad506e1c4a21406f988ef994",
+"libliter_llm_nif-v1.9.0-nif-2.16-x86_64-pc-windows-gnu.dll.tar.gz" => "sha256:12db7feacbb0f6317000bc2700c31c24bd369154ac9d6c01053a3a71a1f69da9",
+"libliter_llm_nif-v1.9.0-nif-2.16-x86_64-unknown-linux-gnu.so.tar.gz" => "sha256:10f7ab4ef1200d645039b6c5c91d5488986182c0be655bdae28da47d5841167c",
+"libliter_llm_nif-v1.9.0-nif-2.17-aarch64-apple-darwin.so.tar.gz" => "sha256:81f22e967bdc06c36ead2198f2f9d891689001d272d3a2fdf30d1340ded60f8a",
+"libliter_llm_nif-v1.9.0-nif-2.17-aarch64-unknown-linux-gnu.so.tar.gz" => "sha256:e7da767291cd4c3f07afca5779ae4e6c7c174a2fc7b6662e4f7b2c1c8355373d",
+"libliter_llm_nif-v1.9.0-nif-2.17-x86_64-pc-windows-gnu.dll.tar.gz" => "sha256:4e7fafb6356b839d9cbfffe7a76f7a84e1f84121e92dafbe5dcc0b26b32c19ac",
+"libliter_llm_nif-v1.9.0-nif-2.17-x86_64-unknown-linux-gnu.so.tar.gz" => "sha256:8395235401094ac18a54114b9a3f27fea36eba82353da8e166c514cb72226865",
 }